+++ ai-logic/firebase-ai/src/main/kotlin/com/google/firebase/ai/generativemodel/CloudGenerativeModelProvider.kt
@@ -121,7 +121,8 @@ internal class CloudGenerativeModelProvider(
     when (generativeBackend.backend) {
       GenerativeBackendEnum.GOOGLE_AI ->
         CountTokensRequest.forGoogleAI(buildGenerateContentRequest(prompt))
-      GenerativeBackendEnum.VERTEX_AI ->
+      GenerativeBackendEnum.VERTEX_AI,
+      GenerativeBackendEnum.AGENT_PLATFORM ->
         CountTokensRequest.forVertexAI(buildGenerateContentRequest(prompt))
     }